legacy/requirements.txt
djangorestframework
django
pydantic[email]
pytest
furl
django-rest-knox
nanoid
python-ulid[pydantic]